+++ src/main/java/com/sonicether/soundphysics/mixin/SourceMixin.java
@@ -34,7 +34,7 @@ public class SourceMixin {
     @ModifyArg(method = "setAttenuation", at = @At(value = "INVOKE", target = "org/lwjgl/openal/AL10.alSourcef (IIF)V", ordinal = 0), index = 2)
     private float AttenuationHijack(int pointer2, int paramid, float attenuation) {
         if (paramid != 4131) throw new IllegalArgumentException("Tried modifying wrong field. No attenuation here.");
-        return attenuation / ConfigManager.getConfig().General.attenuationFactor;
+        return  attenuation / (float)(ConfigManager.getConfig().General.attenuationFactor);
     }
 
 }
